Strukturerat upp Seed, nästa steg är att göra migrate så en seed skapas. Även uppdaterat Prisma Schema.
